Fixed previous token refused issue:
Fixed the problem that detects a refused token when closing the web without logging out and returning to the page where you were. I needed in the PersistLogin.jsx to make only a refresh request. Two were made because in development mode the StrictMode is enabled and the useEffect are mounted twice.