Fixed session not persistence:
When the session was not persisted and the page was refreshed, when returning to the login page if you navigated to a previous url from the address bar manually, the accessToken remained in the global context and you could access home again without logging in.

Now when the session does not persist, the user is logged out, deleting any information from the global context and cookies.

I have also added some comments changes and improvements in the app.